requirements.txt
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
transformers==4.39.3
easyocr==1.7.1
face_recognition==1.3.0
